Extract github proxy helper methods in to their own file
